Move some FTUE related methods and classes back to the `impl` module
